Made functions to calculate input lignin and nitrogen methods of LitterChemistry
